Improve MDX rendering performance
