+++ .changeset/thin-starfishes-love.md
@@ -1,0 +1,5 @@
+---
+'@astrojs/mdx': patch
+---
+
+Improve MDX rendering performance by sharing processor instance
