+++ packages/integrations/mdx/package.json
@@ -74,6 +74,7 @@
     "remark-rehype": "^10.1.0",
     "remark-shiki-twoslash": "^3.1.3",
     "remark-toc": "^8.0.1",
+    "unified": "^10.1.2",
     "vite": "^4.4.9"
   },
   "engines": {
