+++ packages/markdown/remark/src/internal.ts
@@ -1,6 +1,5 @@
 export {
 	InvalidAstroDataError,
 	safelyGetAstroData,
-	setAstroData,
 	toRemarkInitializeAstroData,
 } from './frontmatter-injection.js';
